+++ src/TBureck.BuildingBlocks.Domain/IDomainEvent.cs
@@ -1,0 +1,8 @@
+﻿namespace TBureck.BuildingBlocks.Domain
+{
+    
+    /// <summary>
+    /// Marker interface for domain events. Implement this interface to send/receive events using the Event class.
+    /// </summary>
+    public interface IDomainEvent { }
+}